% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isofor.R
\name{step_isofor}
\alias{step_isofor}
\title{Isolation forest recipe step}
\usage{
step_isofor(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  iso_mod = NULL,
  sample_size = 256,
  num_trees = 100,
  max_depth = 10,
  options = list(nproc = 1, replace = FALSE, seed = 101, respect_unordered_factors =
    "partition"),
  skip = FALSE,
  id = rand_id("isofor")
)
}
\arguments{
\item{recipe}{The recipe object}

\item{...}{Additional arguments}

\item{role}{Role of the transformed output}

\item{trained}{Has the model been trained (TRUE/FALSE)}

\item{iso_mod}{Prepped/trained isolation forest model object.}

\item{sample_size}{(solitude) (positive integer, default = 256) Number of observations in the dataset to used to build a tree in the forest}

\item{num_trees}{(solitude) num_trees: (positive integer, default = 100) Number of trees to be built in the forest}

\item{max_depth}{(solitude) max_depth: (positive number, default: ceiling(log2(sample_size))) See max.depth argument in ranger}

\item{options}{additional default arguments to the solitude isolation forest API.}

\item{skip}{Whether to skip the baking step for new data.}

\item{id}{Name of the step}
}
\value{

}
\description{
Isolation forest recipe step
}
\examples{

library(dplyr)
library(tidymodels)


splits <- initial_split(mtcars)
train <- training(splits)
test <- testing(splits)
resamples <- bootstraps(train, times = 5)

rec_obj <-
recipe(mpg ~ ., data = mtcars) \%>\%
   step_dummy(all_nominal_predictors()) \%>\%
   step_isofor(all_predictors(), sample_size = tune(), max_depth = tune())

lm_mod <- linear_reg() \%>\%
  set_engine("lm")

wf_linear <- workflow() \%>\%
  add_recipe(rec_obj) \%>\%
  add_model(lm_mod)

iso_param <- wf_linear \%>\%
  parameters() \%>\%
  update(sample_size = sample_size(c(1, 24)))

tuned_mod <- wf_linear \%>\%
  tune_grid(resamples = resamples, param_info = iso_param)
}
