% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunable.R
\name{tree_depth}
\alias{tree_depth}
\title{Custom tuning functions for the package.}
\usage{
tree_depth(range = c(1L, 15L), trans = NULL)
}
\arguments{
\item{A}{recipe step object}

\item{...}{Not used}
}
\value{
A tibble
}
\description{
Custom tuning functions for the package.
}
